requirements.txt
asgiref==3.5.2
attrs==22.1.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.0.0
factory-boy==3.2.1
Faker==15.1.3
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
Pillow==9.3.0
pluggy==1.0.0
psycopg2==2.9.5
PyJWT==2.6.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
typing_extensions==4.4.0
